Upgraded to HDRP and switched to Vulkan Graphics API
HDRP gives us way more powerful graphical effects, so I'm hoping we can use it for our more technically difficult effects.

I switched the graphics API to Vulkan purely because OpenGl isn't supported on Linux for HDRP and I develop on linux. Is that a petty reason? Yes, but as long as nothing breaks it's fine.